requirements.txt
google-generativeai
beautifulsoup4
chardet
python-dotenv
